Added Media breakpoints and adaptivity for tablet, mobile. Added css classnames configurations to Vite config. Added Button variants. Separated repeating styles for Button and DatePicker to mixins. Added logo icon in html title. Created Toolbar and AddTodo components and decomposed to separate folders.
