+++ index.html
@@ -6,6 +6,11 @@
       name="viewport"
       content="width=device-width, initial-scale=1.0"
     />
+    <link
+      rel="icon"
+      type="image/svg+xml"
+      href="/logo.svg"
+    />
     <title>TODO app - trainee camp</title>
   </head>
   <body>
