+++ src/components/Toolbar/AddTodo/AddTodo.module.scss
@@ -1,0 +1,4 @@
+.container {
+  display: flex;
+  gap: 20px;
+}
