+++ src/components/Toolbar/Bar/Toolbar.tsx
@@ -1,0 +1,13 @@
+import { FC } from 'react';
+
+import { AddTodo } from '../AddTodo';
+
+import style from './Toolbar.module.scss';
+
+export const Toolbar: FC = () => {
+  return (
+    <section className={style.toolbar}>
+      <AddTodo />
+    </section>
+  );
+};
