+++ src/components/Toolbar/Bar/index.ts
@@ -1,0 +1,1 @@
+export * from './Toolbar';
