+++ src/components/index.ts
@@ -1,6 +1,7 @@
 /* COMMON components */
 export * from './TodoList';
 export * from './TodoForm';
+export * from './Toolbar';
 
 /* UI components */
 export * from './ui/TaskInput';
