+++ src/components/ui/CustomButton/CustomButton.module.scss
@@ -1,26 +1,79 @@
 @import '@style/_variables.scss';
 
-.container {
-  display: flex;
-  padding: 14px;
-
-  border-radius: 10px;
-  border: 1px solid $clr_gray_primary;
-
-  background-color: transparent;
-  fill: $clr_gray_primary;
-  color: $clr_gray_primary;
-
+@mixin with-hover-focus() {
   &:hover {
+    @content;
+  }
+  &:focus {
+    @content;
+  }
+}
+@mixin with-variant($bg_color, $fill_color) {
+  background-color: $bg_color;
+  color: $fill_color;
+  fill: $fill_color;
+}
+
+/* BASE BTN */
+
+.button {
+  display: flex;
+  gap: 10px;
+
+  border-radius: $border_radius;
+  border: 1px solid transparent;
+
+  font-weight: 500;
+
+  &:disabled {
+    opacity: 0.5;
+    pointer-events: none;
+  }
+}
+
+/* SIZES */
+
+.size_default {
+  padding: 5px 15px;
+  max-height: 30px;
+}
+.size_sm {
+  padding: 5px;
+  max-height: 30px;
+}
+.size_md {
+  padding: 15px;
+  max-height: 50px;
+}
+
+/* VARIANTS */
+
+.variant_default {
+  @include with-variant(transparent, black);
+
+  @include with-hover-focus {
+    border: 1px solid black;
+  }
+}
+
+.variant_primary {
+  @include with-variant($clr_blue_primary, white);
+
+  @include with-hover-focus {
     background-color: $clr_blue_secondary;
     fill: $clr_blue_primary;
     color: $clr_blue_primary;
     border: 1px solid $clr_blue_primary;
   }
-  &:focus {
-    background-color: $clr_blue_secondary;
-    fill: $clr_blue_primary;
-    color: $clr_blue_primary;
-    border: 1px solid $clr_blue_primary;
+}
+
+.variant_secondary {
+  @include with-variant($clr_gray_secondary, black);
+
+  @include with-hover-focus {
+    background-color: black;
+    fill: white;
+    color: white;
+    border: 1px solid black;
   }
 }
