+++ src/components/ui/CustomButton/CustomButton.tsx
@@ -5,6 +5,8 @@ import style from './CustomButton.module.scss';
 interface CustomButtonProps {
   children: ReactNode;
   isDisabled?: boolean;
+  size?: 'default' | 'sm' | 'md';
+  variant?: 'default' | 'primary' | 'secondary';
 
   onClick: () => void;
 }
@@ -12,12 +14,18 @@ interface CustomButtonProps {
 export const CustomButton: FC<CustomButtonProps> = ({
   children,
   isDisabled = false,
+  size = 'default',
+  variant = 'default',
 
   onClick,
 }) => {
+  const classes = `${style.button} ${style[`size_${size}`]} ${
+    style[`variant_${variant}`]
+  }`;
+
   return (
     <button
-      className={style.container}
+      className={classes}
       type="button"
       onClick={onClick}
       disabled={isDisabled}
