+++ src/main.tsx
@@ -7,8 +7,6 @@ import { store } from '@redux';
 
 import { App } from './App';
 
-import '@style/Main.scss';
-
 ReactDOM.createRoot(document.getElementById('root')!).render(
   <React.StrictMode>
     <Provider store={store}>
