+++ src/style/App.module.scss
@@ -1,18 +1,38 @@
+@import url('https://fonts.googleapis.com/css2?family=Jost:wght@300;400;500;600;700;800;900&display=swap');
+@import './_reset.scss';
 @import './variables';
+@import './_media';
+
+* {
+  font-family: 'Jost', sans-serif;
+}
 
 .app {
+  max-width: $main_container;
   min-height: 100vh;
+  margin: 0 auto;
+  background-color: $clr_gray_secondary;
 }
 
 .main_container {
   width: 50%;
 
   margin: 0 auto;
-  padding: 50px 0px;
+  padding: 25px 0px;
+
+  display: flex;
+  flex-direction: column;
+  gap: 20px;
 }
 
-.add_task_container {
-  display: flex;
-  gap: 20px;
-  margin-bottom: 20px;
+@include tablet {
+  .main_container {
+    width: 70vw;
+  }
+}
+
+@include mobile {
+  .main_container {
+    width: 90vw;
+  }
 }
