+++ src/style/_media.scss
@@ -1,0 +1,22 @@
+@import './_variables';
+
+@mixin desktop() {
+  @media (max-width: $breakpoint_desktop) {
+    @content;
+  }
+}
+@mixin laptop() {
+  @media (max-width: $breakpoint_laptop) {
+    @content;
+  }
+}
+@mixin tablet() {
+  @media (max-width: $breakpoint_tablet) {
+    @content;
+  }
+}
+@mixin mobile() {
+  @media (max-width: $breakpoint_mobile) {
+    @content;
+  }
+}
