+++ src/style/_reset.scss
@@ -18,8 +18,5 @@
     &:hover {
       cursor: pointer;
     }
-    &:disabled {
-      cursor: not-allowed;
-    }
   }
 }
