--- src/style/main.scss
@@ -1,6 +1,0 @@
-@import url('https://fonts.googleapis.com/css2?family=Jost:wght@300;400;500;600;700;800;900&display=swap');
-@import './_reset.scss';
-
-* {
-  font-family: 'Jost', sans-serif;
-}
